init commit for ReadBytes, a future separate package
ReadBytes currently attempts to mirror the functionality of opening a binary file but for a bytes object. In the future it will also wrap text and binary files. It extends the functionality mostly in providing arbitrarily many file read points, and adding some extra options to reading and writing. When writing to a bytes/file in insert mode, all pointers after the edited bytes are updated to remain on the same byte. Using multiple pointers removes the need for repeated seeking, and can aid in clarity of code. Additionally a ReadBytes class could be maintained outside of, say, a loop, and if was using a file for data, it would keep locations the same while allowing the file to close, as files will be opened and closed on each method call, unless doing so causes a less than negligible runtime penalty, and then it will use file open and close instead of with, and use timed looping seconday processes or some other method to periodically close files if doing so is doable and more user-friendly than open/close methods.