Add compressed image layer data support to file input
not directly tested, as there isn't a working exporter yet. uses ReadBytes (assigned as bytes_filelike, ReadBytes name is not permanent) to handle reading from the uncompressed bytes or to read from the raw image bytes if it is stored uncompressed. I will probably be working on file output next.